mail-processor/pyproject.toml
[project]
name = "mail-processor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "David Birks", email = "david@birks.dev" }
]
dependencies = [
    "langchain>=0.2.5",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.2.5",
    "tavily-python>=0.3.3",
    "langchain-openai>=0.1.8",
    "langgraph>=0.0.69",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mail_processor"]
